chore: rename categories to single noun, update method name and return status
